refactor Music for inheritance
